+++ CHANGELOG.md
@@ -2,4 +2,12 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.2.0](https://github.com/tengre/patterns-in-ts/compare/v0.1.1...v0.2.0) (2022-10-24)
+
+
+### Features
+
+* **custom:** add class Observable ([759a59c](https://github.com/tengre/patterns-in-ts/commit/759a59ce6e0c3e9f1ee288a2eb5e6ecdee2d416d))
+* **custom:** add Observer class ([5d7a506](https://github.com/tengre/patterns-in-ts/commit/5d7a506cad05be4a6d8997d9ec4097666637eb3f))
+
 ### 0.1.1 (2022-10-24)